requirements.txt
requests
selenium
webdriver_manager
fake_useragent
stem
tqdm
pyshortcuts
pyperclip
Pillow